refactor lab runner preparation tests
